requirements.txt
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
urllib3<2.0
beautifulsoup4